\name{icmm-package}
\alias{icmm-package}
\docType{package}
\title{Empirical Bayes Variable Selection via ICM/M}
\description{Carries out empirical Bayes variable selection via ICM/M algorithm. The basic problem is to fit a high-dimensional regression which most of the coefficients are assumed to be zero. This package allows incorporating the Ising prior to capture structure of predictors in the modeling process. The current version of this package can handle the normal, binary logistic, and Cox's regression.}
\details{
\tabular{ll}{
Package: \tab icmm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2017-07-24\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Vitara Pungpapong, Min Zhang, Dabao Zhang

Maintainer: Vitara Pungpapong <vitara@cbs.chula.ac.th>
}
\references{
Pungpapong, V., Zhang, M. and Zhang, D. (2015). Selecting massive variables using an iterated conditional modes/medians algorithm. Electronic Journal of Statistics. 9:1243-1266. <doi:10.1214/15-EJS1034>.\cr
Pungpapong, V., Zhang, M. and Zhang, D. (2017). Variable selection for high-dimensional generalized linear models using an iterated conditional modes/medians algorithm. Preprint <arXiv:1707.08298>.
}
