\name{initbetaBinomial}
\alias{initbetaBinomial}
\docType{data}
\title{Initial values for the regression coefficients used in example for running ICM/M algorithm in binary logistic model}
\description{Initial values for the regression coefficients obtained from binary logistic model with lasso regularization for \code{simBinomial} data set.}
\usage{data(initbetaBinomial)}
\format{
  A data frame with 400 rows.
  \describe{
    \item{\code{V1}}{a numeric vector of the regression coefficients.}
  }
}
\examples{
  data(initbetaBinomial)
}

