\name{initbetaCox}
\alias{initbetaCox}
\docType{data}
\title{Initial values for the regression coefficients used in example for running ICM/M algorithm in Cox's model}
\description{Initial values for the regression coefficients obtained from Cox's model with lasso regularization for \code{simCox} data set.}
\usage{data(initbetaCox)}
\format{
  A data frame with 400 rows.
  \describe{
    \item{\code{V1}}{a numeric vector of the regression coefficients.}
  }
}
\examples{
  data(initbetaCox)
}

