\name{initbetaGaussian}
\alias{initbetaGaussian}
\docType{data}
\title{Initial values for the regression coefficients used in example for running ICM/M algorithm in normal linear regression  model}
\description{Initial values for the regression coefficients obtained from normal linear regression model with lasso regularization for \code{simGaussian} data set.}
\usage{data(initbetaGaussian)}
\format{
  A data frame with 400 rows.
  \describe{
    \item{\code{V1}}{a numeric vector of the regression coefficients.}
  }
}
\examples{
  data(initbetaGaussian)
}

